🔀 move user loading logic from AuthService to MemberService
