--- src/main/java/com/grepp/diary/app/model/ai/AiRepository.java
@@ -1,8 +1,0 @@
-package com.grepp.diary.app.model.ai;
-
-import com.grepp.diary.app.model.ai.entity.Ai;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AiRepository extends JpaRepository<Ai, Integer> {
-
-}
